Создает контроллер, переносит всю логику по рендеру точек путешествия в него
